+++ app/src/main/java/com/example/simplemoviecatalog/ui/activities/DetailActivity.kt
@@ -7,7 +7,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.isVisible
 import com.example.simplemoviecatalog.databinding.ActivityDetailBinding
 import com.example.simplemoviecatalog.domain.model.DomainFavoritesModel
-import com.example.simplemoviecatalog.domain.model.DomainModel
 import com.example.simplemoviecatalog.ui.adapters.favorite.FavoritesAdapter
 import com.example.simplemoviecatalog.ui.viewModels.FavoriteViewModel
 import com.example.simplemoviecatalog.utils.Constants
@@ -22,6 +21,7 @@ class DetailActivity : AppCompatActivity() {
     private val favoritesAdapter: FavoritesAdapter by lazy {
         FavoritesAdapter(emptyList())
     }
+    private lateinit var movie: DomainFavoritesModel
 
     private val favoritesViewModel: FavoriteViewModel by viewModels()
 
@@ -30,6 +30,7 @@ class DetailActivity : AppCompatActivity() {
         binding = ActivityDetailBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        movie =  intent.extras?.getParcelable("movie")!!
         getMovies()
         binding.btnAddToFavorites.setOnClickListener { verifyFavorites() }
         binding.btnDeleteToFavorites.setOnClickListener { deleteFromFavorites() }
@@ -37,8 +38,7 @@ class DetailActivity : AppCompatActivity() {
     }
     //Recupero los datos del MainActivity
     private fun getMovies() {
-        val movie = intent.extras?.getParcelable<DomainModel>("movie")
-        movie?.let {
+        movie.let {
             binding.tvTitle.text = it.title
             binding.tvOverview.text = it.overview
             binding.tvReleaseDate.text = it.releaseDate
@@ -49,8 +49,7 @@ class DetailActivity : AppCompatActivity() {
     }
     //Verifica si la pelicula ya se encuentra en la base de datos y desabilita el boton
     private fun verifyFavorites() {
-        val movie = intent.extras?.getParcelable<DomainModel>("movie")
-        movie?.let {
+        movie.let {
             favoritesViewModel.verifyFavorite(it.title)
             favoritesViewModel.verifyLiveData.observe(this) { isFavorite ->
                 if (isFavorite) {
@@ -66,8 +65,7 @@ class DetailActivity : AppCompatActivity() {
     }
 
 
-    private fun addToFavorites(movie:DomainModel) {
-        val movie = createMovieInstance(movie)
+    private fun addToFavorites(movie:DomainFavoritesModel) {
         favoritesViewModel.addToFavorites(movie)
         favoritesViewModel.getFavorites()
         Toast.makeText(this, "Se ha agregado a favoritos.",
@@ -77,13 +75,11 @@ class DetailActivity : AppCompatActivity() {
 
     //deberia eliminar y actualizar la lista de favoritos
     private fun deleteFromFavorites() {
-        val movie = intent.extras?.getParcelable<DomainModel>("movie")
-        movie?.let {
-            val favoritesMovie = createMovieInstance(movie)
+        movie.let {
             //Borra una pelicula de la base de datos
-            favoritesViewModel.deleteFavoriteMovie(favoritesMovie)
+            favoritesViewModel.deleteFavoriteMovie(movie)
             //Saca y actualiza el adapter...supuestamente
-            favoritesAdapter.removeFavorite(favoritesMovie)
+            favoritesAdapter.removeFavorite(movie)
             Toast.makeText(this, "Se ha eliminado de favoritos.", Toast.LENGTH_SHORT).show()
             //Al eliminar una pelicula, desabilita el boton
             binding.btnAddToFavorites.isEnabled = true
@@ -91,13 +87,4 @@ class DetailActivity : AppCompatActivity() {
         }
     }
 
-    private fun createMovieInstance(movie: DomainModel): DomainFavoritesModel {
-        return DomainFavoritesModel(
-            title = movie.title,
-            overview = movie.overview,
-            releaseDate = movie.releaseDate,
-            voteAverage = movie.voteAverage,
-            image = movie.image
-        )
-    }
 }
