+++ app/src/main/java/com/example/simplemoviecatalog/ui/adapters/movie/PopularMoviesAdapter.kt
@@ -7,21 +7,13 @@ import android.widget.Filterable
 import androidx.recyclerview.widget.RecyclerView
 import com.example.simplemoviecatalog.R
 import com.example.simplemoviecatalog.domain.model.DomainModel
-import com.example.simplemoviecatalog.ui.adapters.MoviesListFilter
 
 class PopularMoviesAdapter(
     private var popularMoviesList: List<DomainModel>,
     private var onClickMoviesListener: OnClickMoviesListener?
 ) : RecyclerView.Adapter<PopularMoviesViewHolder>(), Filterable {
 
-
-    private val filter = MoviesListFilter(this)
     var filteredPopularMoviesList: List<DomainModel> = emptyList()
-
-    init {
-        filteredPopularMoviesList = popularMoviesList
-    }
-
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PopularMoviesViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
@@ -48,6 +40,25 @@ class PopularMoviesAdapter(
         notifyDataSetChanged()
     }
 
-    override fun getFilter(): Filter = filter
+    override fun getFilter(): Filter {
+        return object : Filter() {
+            override fun performFiltering(constraint: CharSequence?): FilterResults {
+                val filteredList = if (constraint.isNullOrEmpty()) {
+                    popularMoviesList
+                } else {
+                    popularMoviesList.filter { it.title.contains(constraint, true) }
+                }
+                val filterResults = FilterResults()
+                filterResults.values = filteredList
+                return filterResults
+            }
 
+            override fun publishResults(constraint: CharSequence?, results: FilterResults?) {
+                @Suppress("UNCHECKED_CAST")
+                filteredPopularMoviesList = results?.values as List<DomainModel>?
+                    ?: emptyList()
+                notifyDataSetChanged()
+            }
+        }
+    }
 }
